fatigue_detection/requirements.txt
numpy>=1.26
opencv-python>=4.10
mediapipe>=0.10
scikit-learn>=1.5
torch>=2.4
onnx>=1.17
onnxruntime>=1.20
Pillow>=10.0
